% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_button.R
\name{download_button}
\alias{download_button}
\title{Custom learnr download button}
\usage{
download_button(df, filename = "data.csv", buttonLabel = "Download CSV")
}
\arguments{
\item{df}{a data frame object}

\item{filename}{name of the file -- include the ".csv" in the name}

\item{buttonLabel}{label of the button}
}
\value{
a button
}
\description{
Custom learnr download button
}
