% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envx-package.R
\docType{package}
\name{envx-package}
\alias{envx}
\alias{envx-package}
\title{envx: ENVX: A package for ENVX Units at the University of Sydney}
\description{
A package that contains tutorials, functions and data for use in ENVX1002 and ENVX2001 Units at the University of Sydney.
}
\author{
\strong{Maintainer}: Januar Harianto \email{januar.harianto@sydney.edu.au}

}
\keyword{internal}
