% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_tutorials.R
\name{list_tutorials}
\alias{list_tutorials}
\title{List available tutorials and open if asked}
\usage{
list_tutorials()
}
\value{
A list of available tutorials, and opens the selected tutorial if
requested
}
\description{
List available tutorials and open if asked
}
